consolidate routes, formatting, minor renames
